requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
easyocr==1.7.2
filelock==3.20.2
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
groq==1.0.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
importlib-metadata==6.11.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.13.0
networkx==3.6.1
ninja==1.13.0
numpy==1.26.4
opencv-python==4.8.1.78
opencv-python-headless==4.12.0.88
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
proto-plus==1.27.0
protobuf==4.25.8
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.1
pytesseract==0.3.10
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
scikit-image==0.26.0
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.29.0
sympy==1.14.0
tenacity==8.5.0
tifffile==2025.12.20
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
validators==0.35.0
watchdog==6.0.0
zipp==3.23.0
